+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # OldNews ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- The new subscription dialog now defaults the feed to any URL found in the
+  clipboard. ([#119](https://github.com/davep/oldnews/pull/119))
+
 ## v0.9.0
 
 **Released: 2026-02-06**